Refactored image-uploader to the API
