--- api/v1/blog/index.js
@@ -1,1 +1,0 @@
-module.exports.routes = require('./routes')
