+++ api/v1/imageUploader/imageUploader.routes.js
@@ -1,0 +1,13 @@
+import express from 'express'
+import {
+    getImageUploader,
+    postImageUploader
+} from './imageUploader.controller'
+import {upload} from './upload'
+
+const router = express.Router()
+
+router.get('/', getImageUploader)
+router.post('/', upload.array('image', 1), postImageUploader);
+
+module.exports = router
