+++ api/v1/index.js
@@ -1,13 +1,15 @@
-var blog = require('./blog')
-var contactRoutes = require('./contact/contact.routes')
-var github = require('./github/github.routes')
+import express from 'express'
+import blog from './blog/blog.routes'
+import contact from './contact/contact.routes'
+import github from './github/github.routes'
+import imageUploader from './imageUploader/imageUploader.routes'
+import {isAuthenticated} from '../../web/pages/admin/is-authenticated'
 
-var express = require('express')
-var router = express()
+const router = express()
 
-// API Route
-router.use('/contact', contactRoutes)
-router.use('/blog', blog.routes)
+router.use('/blog', blog)
+router.use('/contact', contact)
 router.use('/github', github)
+router.use('/upload-image', isAuthenticated, imageUploader)
 
 module.exports = router
