+++ web/pages/admin/admin.routes.js
@@ -2,7 +2,6 @@ import express from 'express'
 import passport from 'passport'
 import Recaptcha from 'express-recaptcha'
 
-import imageUploader from './imageUploader'
 import seriesRouter from './admin.series.routes'
 import env from '../../../env'
 import {isAuthenticated} from './is-authenticated'
@@ -36,7 +35,6 @@ router.get('/delete/:permalink', isAuthenticated, deletePostConfirmation)
 router.get('/delete/:permalink/confirm', isAuthenticated, deletePost)
 
 router.use('/series', isAuthenticated, seriesRouter)
-router.use('/uploader', isAuthenticated, imageUploader)
 
 // General access routes
 router.get('/logout', logout)
